Add loading/error feedback, match details link
Show an animated progress bar while match data is
loading. Afterward, either show the match graphs
or an error section with some possible causes and
suggested steps.

Since one of the suggested steps is to try jumping
to the official match details from Riot, we need
better region constants. This is pretty useful
stuff to have, so include a match details link
above the graphs too.